requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
discord.py==2.5.2
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.2.1
numpy==2.0.2
#nvidia-cublas-cu12==12.6.4.1
#nvidia-cuda-cupti-cu12==12.6.80
#nvidia-cuda-nvrtc-cu12==12.6.77
#nvidia-cuda-runtime-cu12==12.6.77
#nvidia-cudnn-cu12==9.5.1.17
#nvidia-cufft-cu12==11.3.0.4
#nvidia-cufile-cu12==1.11.1.6
#nvidia-curand-cu12==10.3.7.77
#nvidia-cusolver-cu12==11.7.1.2
#nvidia-cusparse-cu12==12.5.4.2
#nvidia-cusparselt-cu12==0.6.3
#nvidia-nccl-cu12==2.26.2
#nvidia-nvjitlink-cu12==12.6.85
#nvidia-nvtx-cu12==12.6.77
openai==1.93.0
packaging==25.0
pdfminer==20191125
pdfminer.six==20250506
pillow==11.3.0
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.0.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
#triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
